+++ docs/changelog.md
@@ -10,7 +10,8 @@ Version 0.0.16
 MML changes:
  * Fixed a note tracking bug that can cause successive `{}` portamentos to not key-on.
    For example, `{cd} {cf} {cg}` will only play a single portamento.
- * Added the `?@` subroutine instrument hint command
+ * Added the `?@` subroutine instrument hint command.
+ * Fixed a state tracking bug involving loops that can cause dropped instructions.
 
 GUI changes:
  * Sample Analyser can now analyse samples up to a length of 32768.